Bot api integration

+++ BettingShop.TelegramBot/Executor/Executors/CreateEventExecutor.cs
@@ -1,5 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
+using BettingShop.Api.Client;
+using BettingShop.Api.Client.Models;
 using BettingShop.TelegramBot.Command.Commands;
 using BettingShop.TelegramBot.Message;
 using Telegram.Bot;
@@ -21,29 +23,67 @@ namespace BettingShop.TelegramBot.Executor.Executors
         public async Task ExecuteAsync(UserMessage message)
         {
             var state = stateService.GetCurrentState(message.User);
+            var eventClient = new BetEventClient("http://localhost:27254");
             if (state is CreateEventCommandState createState)
             {
                 switch (createState.State)
                 {
                     case CreateEventState.Name:
+                        if (message.Tail == "")
+                        {
+                            await client.SendTextMessageAsync(message.TelegramMessage.Chat, "Введите название события");
+                            stateService.SaveState(message.User, new CreateEventCommandState() { State = CreateEventState.Name });
+                            break;
+                        }
                         await client.SendTextMessageAsync(message.TelegramMessage.Chat,
-                            $"Введите возможные исходы вашего события");
-                        stateService.SaveState(message.User, new CreateEventCommandState() { State = CreateEventState.Outcomes });
+                            $"Введите возможные исходы вашего события, каждое с новой строки");
+                        stateService.SaveState(message.User, new CreateEventCommandState() { State = CreateEventState.Outcomes, Name = message.Tail });
 
                         break;
                     case CreateEventState.Outcomes:
-                        await client.SendTextMessageAsync(message.TelegramMessage.Chat, 
+                        if (message.Tail == "")
+                        {
+                            await client.SendTextMessageAsync(message.TelegramMessage.Chat, "Введите название события");
+                            stateService.SaveState(message.User, new CreateEventCommandState() { State = CreateEventState.Name });
+                            break;
+                        }
+                        if (message.Tail.IndexOf('\n') == -1)
+                        {
+                            await client.SendTextMessageAsync(message.TelegramMessage.Chat,
+                                $"Введите, пожалуйста, несколько исходов, каждый с новой строки");
+                            break;
+                        }
+                        await client.SendTextMessageAsync(message.TelegramMessage.Chat,
                             $"Введите дату и время окончания события");
-                        stateService.SaveState(message.User, new CreateEventCommandState() { State = CreateEventState.Deadline });
+                        stateService.SaveState(message.User, new CreateEventCommandState() { State = CreateEventState.Deadline, Name = createState.Name, Outcomes = message.Tail });
                         break;
                     case CreateEventState.Deadline:
-                        var date = DateTime.Parse(message.Tail);
+                        DateTime date;
+                        if (!DateTime.TryParse(message.Tail, out date))
+                        {
+                            await client.SendTextMessageAsync(message.TelegramMessage.Chat,
+                                $"Введите существующую дату");
+                            break;
+                        }
+
+                        if (date < DateTime.Now)
+                        {
+                            await client.SendTextMessageAsync(message.TelegramMessage.Chat,
+                                $"Введите дату в будущем");
+                            break;
+                        }
                         await client.SendTextMessageAsync(message.TelegramMessage.Chat,
                             $"Напишите описание своего события, если не хотите, поставьте \"-\"");
-                        stateService.SaveState(message.User, new CreateEventCommandState() { State = CreateEventState.Description });
+                        stateService.SaveState(message.User, new CreateEventCommandState() { State = CreateEventState.Description, Name = createState.Name, Outcomes = createState.Outcomes, Deadline = date });
                         break;
                     case CreateEventState.Description:
-                        //сгенерить событие
+                        if (message.Tail == "")
+                        {
+                            await client.SendTextMessageAsync(message.TelegramMessage.Chat, "Введите название события");
+                            stateService.SaveState(message.User, new CreateEventCommandState() { State = CreateEventState.Name });
+                            break;
+                        }
+                        await eventClient.CreateAsync(new BetEventMeta { Name = createState.Name, BetDeadline = createState.Deadline, Outcomes = createState.Outcomes, Description = message.Tail, CreatorId = message.TelegramMessage.From.Id });
                         await client.SendTextMessageAsync(message.TelegramMessage.Chat,
                             $"Ваше событие сохранено");
                         stateService.DeleteState(message.User);
@@ -55,9 +95,6 @@ namespace BettingShop.TelegramBot.Executor.Executors
                 await client.SendTextMessageAsync(message.TelegramMessage.Chat, "Введите название события");
                 stateService.SaveState(message.User, new CreateEventCommandState() { State = CreateEventState.Name });
             }
-
-            await client.SendTextMessageAsync(message.TelegramMessage.Chat, $"Called CreateEventExecutor with message {message.TelegramMessage.Text} " +
-                                                                           $"and state {stateService.GetCurrentState(message.User)}");
         }
 
     }
